backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
passlib[bcrypt]
pyjwt
pymongo
email-validator
python-dotenv
requests
openai
itsdangerous
slowapi